moved particles_data_gpu next to particles_data
