+++ sph/include/sph/cuda/CMakeLists.txt
@@ -1,9 +1,9 @@
 if(CMAKE_HIP_COMPILER)
-    set_source_files_properties(cuda_density.cu cuda_iad.cu cuda_grad_p.cu gpu_particle_data.cu PROPERTIES LANGUAGE HIP)
+    set_source_files_properties(cuda_density.cu cuda_iad.cu cuda_grad_p.cu ../particles_data_gpu.cu PROPERTIES LANGUAGE HIP)
 endif()
 
 if(CMAKE_CUDA_COMPILER OR CMAKE_HIP_COMPILER)
-    add_library(cuda_sph OBJECT cuda_density.cu cuda_iad.cu cuda_grad_p.cu gpu_particle_data.cu)
+    add_library(cuda_sph OBJECT cuda_density.cu cuda_iad.cu cuda_grad_p.cu ../particles_data_gpu.cu)
     target_include_directories(cuda_sph PRIVATE ${PROJECT_SOURCE_DIR}/include)
     target_include_directories(cuda_sph PRIVATE ${CSTONE_DIR})
     target_compile_definitions(cuda_sph PRIVATE USE_CUDA)
