+++ sph/include/sph/cuda/cuda_density.cu
@@ -1,6 +1,6 @@
 #include <algorithm>
 
-#include "sph.cuh"
+#include "sph/sph.cuh"
 #include "sph/particles_data.hpp"
 #include "sph/util/cuda_utils.cuh"
 #include "sph/hydro_3L/density_kern.hpp"
