+++ sph/include/sph/hydro_3L/density.hpp
@@ -6,7 +6,7 @@
 
 #include "density_kern.hpp"
 #ifdef USE_CUDA
-#include "sph/cuda/sph.cuh"
+#include "sph/sph.cuh"
 #endif
 
 namespace sphexa
