+++ sph/include/sph/hydro_3L/iad.hpp
@@ -4,7 +4,7 @@
 #include "sph/kernels.hpp"
 #include "iad_kern.hpp"
 #ifdef USE_CUDA
-#include "sph/cuda/sph.cuh"
+#include "sph/sph.cuh"
 #endif
 
 namespace sphexa
