+++ sph/include/sph/hydro_ve/av_switches.hpp
@@ -33,7 +33,7 @@
 
 #include "av_switches_kern.hpp"
 #ifdef USE_CUDA
-#include "sph/cuda/sph.cuh"
+#include "sph/sph.cuh"
 #endif
 
 namespace sphexa
