+++ sph/include/sph/hydro_ve/iad_divv_curlv.hpp
@@ -34,7 +34,7 @@
 #include "iad_kern.hpp"
 #include "divv_curlv_kern.hpp"
 #ifdef USE_CUDA
-#include "sph/cuda/sph.cuh"
+#include "sph/sph.cuh"
 #endif
 
 namespace sphexa
