+++ sph/include/sph/hydro_ve/momentum_energy.hpp
@@ -33,7 +33,7 @@
 
 #include "momentum_energy_kern.hpp"
 #ifdef USE_CUDA
-#include "sph/cuda/sph.cuh"
+#include "sph/sph.cuh"
 #endif
 
 namespace sphexa
