+++ sph/include/sph/hydro_ve/ve_norm_gradh.hpp
@@ -33,7 +33,7 @@
 
 #include "ve_norm_gradh_kern.hpp"
 #ifdef USE_CUDA
-#include "sph/cuda/sph.cuh"
+#include "sph/sph.cuh"
 #endif
 
 namespace sphexa
